修正时间精度问题

+++ jsonparse.go
@@ -689,7 +689,7 @@ func Value2Json(v *DxValue,escapestr bool, dst []byte)[]byte  {
 		dst = strconv.AppendInt(dst,v.Int(),10)
 	case VT_DateTime:
 		dst = append(dst,"\"/Date("...)
-		unixs := int64((DxCommonLib.TDateTime)(v.Float()).ToTime().Unix()*1000)
+		unixs := int64((DxCommonLib.TDateTime)(v.Double()).ToTime().Unix()*1000)
 		dst = strconv.AppendInt(dst,unixs,10)
 		dst = append(dst,")/\""...)
 	case VT_NULL:
